src/backend/requirements.txt
chess==1.8.0
elasticsearch==8.2.2
google-api-python-client==2.39.0
grpc-interceptor==0.14.1
grpcio==1.44.0
grpcio-reflection==1.44.0
grpcio-tools==1.44.0
loguru==0.6.0
opentelemetry-api==1.11.1
opentelemetry-exporter-otlp-proto-grpc==1.11.1
opentelemetry-instrumentation-grpc==0.30b1
opentelemetry-sdk==1.11.1
python-dotenv==0.20.0
structlog==21.5.0
